Combined the different intake and outtake commands into one to make it more simplictly.
